refactor: Move the cpp code to the common directory.
