+++ common/cpp/uuidxx/uuidxx.cc
