+++ common/cpp/uuidxx/uuidxx.h
